requirements.txt
Django==1.8.4
djangorestframework==3.3.2
django-rest-framework-social-oauth2==1.0.4
slugify==0.0.1
django-activity-stream==0.6.3
whitenoise==3.2.2